Updated main to allow for more indepth testing
Main has been updated to allow for more in depth testing.  Also, I checked the delta functions and fixed minor indexing errors in them.  I added gradient descent.  It is not working, not sure what is going wrong at this time.  It seems that the neural network is always trying to return the same value regardless of input.